pyproject.toml
[project]
name = "fraggle"
version = "0.1.0"
description = "A simple RAG API - build a question and answer interface to your own content in minutes"
readme = "README.md"
requires-python = ">=3.12"
authors = [{name = "Tom Dyson", email = "tom@torchbox.com"}]
license = {text = "MIT"}
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "faiss-cpu>=1.8.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.1.0",
    "typer>=0.12.0",
    "any-llm-sdk[openai,anthropic]>=0.1.0",
    "tiktoken>=0.7.0",
]

[project.scripts]
fraggle = "fraggle:cli_wrapper"

[project.urls]
Home = "https://github.com/tomdyson/fraggle"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
